requirements.txt
flask
twilio
requests
python-dotenv
openai
google-generativeai
psycopg2-binary
gunicorn
